+++ src/main/java/de/maxhenkel/corpse/Config.java
@@ -28,26 +28,34 @@ public class Config {
         public ForgeConfigSpec.IntValue corpseForceDespawnTime;
         public ForgeConfigSpec.IntValue corpseSkeletonTime;
         public ForgeConfigSpec.BooleanValue spawnCorpseOnFace;
+        public ForgeConfigSpec.BooleanValue fallIntoVoid;
+        public ForgeConfigSpec.BooleanValue lavaDamage;
 
         public ServerConfig(ForgeConfigSpec.Builder builder) {
             onlyOwnerAccess = builder
                     .comment("If only the owner of the corpse can access the inventory")
-                    .define("only_owner_access", false);
+                    .define("corpse.access.only_owner", false);
             skeletonAccess = builder
                     .comment("If everybody can access the corpse in the skeleton stage (Only if only_owner_access is set to true)")
-                    .define("skeleton_access", false);
+                    .define("corpse.access.skeleton", false);
             corpseDespawnTime = builder
                     .comment("The time passed after a corpse despawns (only if empty)")
-                    .defineInRange("corpse_despawn_time", 20 * 30, 20, Integer.MAX_VALUE);
+                    .defineInRange("corpse.despawn.time", 20 * 30, 20, Integer.MAX_VALUE);
             corpseForceDespawnTime = builder
                     .comment("The time passed after a corpse despawns even if its not empty (-1 = never)")
-                    .defineInRange("corpse_force_despawn_time", -1, -1, Integer.MAX_VALUE);
+                    .defineInRange("corpse.despawn.force_time", -1, -1, Integer.MAX_VALUE);
             corpseSkeletonTime = builder
                     .comment("The time passed after a corpse turns into a skeleton")
-                    .defineInRange("corpse_skeleton_time", 20 * 60 * 60, 0, Integer.MAX_VALUE);
+                    .defineInRange("corpse.skeleton_time", 20 * 60 * 60, 0, Integer.MAX_VALUE);
             spawnCorpseOnFace = builder
-                    .comment("If the corpse should spawn on its face")
-                    .define("spawn_corpse_on_face", false);
+                    .comment("If the corpse should lie on its face")
+                    .define("corpse.lie_on_face", false);
+            fallIntoVoid = builder
+                    .comment("If the corpse should fall into the void")
+                    .define("corpse.fall_into_void", false);
+            lavaDamage = builder
+                    .comment("If the corpse should get removed when in lava")
+                    .define("corpse.lava_damage", false);
         }
     }